Remove rbytes dependency
It doesn't really build anymore (node-waf) and is pointless with the crypto
module. I'm a little sketched by sockjs using Math.random when it's unavailable
(wtf...), but inspecting the source, it's not actually used by anything. (It's
returned as a key in the /info URL, but the key doesn't seem used anywhere.)